[MIPS] Fix initialization of audio mixer track resources.
The value 1 << 32 (maxNumTracks defaults to 32) is undefined by C,
and differs on MIPS than the other arch.

Therefore the track resources were not initialized properly, resulting
in failure to play any audio. This caused all media playback to fail,
including thru Gallery App and the browsers.

The fix allows the mConfiguredNames bitmask to be correctly set to all
1's in the 32-track case, for all arch.

Change-Id: If28d8eff3e2baa45a10c0ff1b42efab1ce7fe693